+++ CHANGELOG.md
@@ -2,5 +2,15 @@
 
 ## Unreleased
 
+- (bump) use cosmos-sdk package v0.47.11
+- (bump) replace github.com/tendermint/tendermint by github.com/cometbft/cometbft v0.37.5
+- update example folder
+
+## [v1.0.1]
+
+- Move example folder
+
+## [v1.0.0]
+
 - Implemented the first version of the client package
 - Implemented the first version of the requester package